Introduction of a personal intelligence workload
This demonstration inferences a llama2 model on our hardware. The
personal intelligence is (not particularly) intelligent, just yet.

**A note to reviewers.** Please take a look at the packaging model. I
really do want to stay near and dear to the Python community vibe.
I am not opposed to alternate packaging approaches. Any software
within the runtime really does need to be packaged using whatever
approach the ecosystem we integrate with typically uses. For this
example, the code is Python, so we need to use **"Pythonic"**
packaging. Brew, Nix, Portage, etc, these are all alternatives.
We need to pick our **Pythonic** packaging and ensure any workloads
use that packaging model.

My proposal for Pythonic packaging is realized in this PR. I also
propose we do not hedge.

Cage match?

Alternative packaging may be a distraction.

Alternative packaging may generate demand... Therefore, I propose
Nix (or Brew) should be documented, if the overhead vs demand generation
is in our favor.