Build pytorch
We have tried many approaches to building, and this seems the most
sound. I have managed 150 container images as part of OpenStack Kolla
and that was pretty straightforward.

The buildx builder has added a feature `output` which will output
build time results from `docker buildx` to `${PWD}/target`.

This is good enough for now.