Produce a knowledge base index
The puprose of this PR is to provide a workload that can be built
into a container. Then the container can be run to build the knowledge base.

The knowledge base indexing is modeled using a worklfow. The workflow contains
serial steps numbered from 0000..00005. This PR isn't quite complete, specifically
the paths are not set properly in all cases. Additionally the volume mounts needed
to mount the host filesystem (to access the knowledge base jsonl file) is not
completed.