Tune the bookworm defaults
Divide the configuration into three sectiions
Set the default image size to 1250G
Output the image in a target path that our environment understands.

Our environment provides a virtiofd device with the tag
storage_bindmount_0000. We will use this device tag to bind the path
`/mnt/repos/targets` to the `storage_bindmount_0000`. This will
provide host-specific storage. This could be improved in the future to
virtualize over a network device (such as iscsi), alhtough this is
GEFN.